+++ rapier/tests/body_types.rs
@@ -9,6 +9,7 @@ use bevy::reflect::TypeRegistryArc;
 
 use heron_core::{Body, BodyType};
 use heron_rapier::rapier::dynamics::{IntegrationParameters, RigidBodySet};
+use heron_rapier::rapier::geometry::ColliderSet;
 use heron_rapier::{BodyHandle, RapierPlugin};
 
 fn test_app() -> App {
@@ -64,6 +65,26 @@ fn create_static_body() {
 }
 
 #[test]
+fn create_sensor_body() {
+    let mut app = test_app();
+
+    let entity = app.world.spawn((
+        GlobalTransform::default(),
+        Body::Sphere { radius: 10.0 },
+        BodyType::Sensor,
+    ));
+
+    app.update();
+
+    let colliders = app.resources.get::<ColliderSet>().unwrap();
+    let body = colliders
+        .get(app.world.get::<BodyHandle>(entity).unwrap().collider())
+        .unwrap();
+
+    assert!(body.is_sensor())
+}
+
+#[test]
 fn can_change_to_static_after_creation() {
     let mut app = test_app();
 
@@ -84,6 +105,30 @@ fn can_change_to_static_after_creation() {
             .unwrap();
 
         assert!(body.is_static());
+    }
+}
+
+#[test]
+fn can_change_to_sensor_after_creation() {
+    let mut app = test_app();
+
+    let entity = app
+        .world
+        .spawn((GlobalTransform::default(), Body::Sphere { radius: 10.0 }));
+
+    app.update();
+
+    app.world.insert_one(entity, BodyType::Sensor).unwrap();
+
+    app.update();
+
+    {
+        let colliders = app.resources.get::<ColliderSet>().unwrap();
+        let collider = colliders
+            .get(app.world.get::<BodyHandle>(entity).unwrap().collider())
+            .unwrap();
+
+        assert!(collider.is_sensor());
     }
 }